+++ changelog.md
@@ -1,3 +1,12 @@
+## 2.5.6
+
+#### Fixed
+User email is trimmed of whitespaces before further processing now
+#### Changes
+Added "MIR" cards support for payments via Google Pay
+Added payment via Tinkoff Pay
+#### Additions
+
 ## 2.5.5
 
 #### Fixed